CategoryCreationScreen -> MVVM
